backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
pymysql==1.1.0
Flask-Cors==4.0.0
requests==2.31.0
cryptography==46.0.3
werkzeug==3.1.5
PyJWT==2.8.0
Pillow==10.2.0
Flask-Limiter==3.5.0
